work_in_progress: Fix progressive scrolling.
todo: Figure out the correct parameters to pass to
`update_widget.update_padding()` in order to make progressive
scrolling magically seamless.
todo: Restore test coverage.

To be clear: We intend to squash this commit, the commit that actually
implements the two section layout and the prep commit that breaks
progressive scrolling together.