buddy_list: Rename references to "keys", "items" and "user_ids".
This is a prep commit for enabling a two section layout in the
buddy_list. The sections will be for "users" and "others".

In this commit we rename:
"keys" to "user_keys",
"items" to "user_items",
and modify buddy_list.populate to expect "key_groups" rather than
"{keys: user_ids}" as well as make a similar change wherever it is
called.