buddy_list: Implement two section layout.
This commit implements a new feature which splits the buddy_list into
two sections, which the code internally refers to as the "users"
section and the "others" section, but for which we conditionally
display the titles "Subscribers"/"Recipients" and "Others".

This commit has several comments about changes that still need to be
made.

This commit adds a class ListCursorWithScrollOverride which, as the
name suggests, overrides the adjust_scroll logic, this is necessary
because the addition of a "pos: sticky" heading element *within* the
scroll container breaks our calculation for scroll_deltas, the use of
scrollIntoView is still buggy when we scroll upwards, just in a
different way. However, it is comparatively better.